+++ frozen-collections-core/src/fz_sets/fz_string_set.rs
@@ -5,8 +5,9 @@ use crate::sets::decl_macros::{debug_trait_funcs, partial_eq_trait_funcs, set_qu
 use crate::sets::{IntoIter, Iter};
 use crate::traits::{Len, Set, SetExtras, SetIteration, SetOps, SetQuery};
 use core::fmt::Debug;
-use core::hash::BuildHasher;
+use core::hash::{BuildHasher, Hash};
 use core::ops::{BitAnd, BitOr, BitXor, Sub};
+use equivalent::Equivalent;
 
 #[cfg(not(feature = "std"))]
 use alloc::vec::Vec;
@@ -29,11 +30,11 @@ use alloc::boxed::Box;
 #[doc = include_str!("../doc_snippets/about.md")]
 #[doc = include_str!("../doc_snippets/hash_warning.md")]
 #[derive(Clone)]
-pub struct FzStringSet<BH = DefaultBuildHasher> {
-    map: FzStringMap<(), BH>,
+pub struct FzStringSet<T, BH = DefaultBuildHasher> {
+    map: FzStringMap<T, (), BH>,
 }
 
-impl FzStringSet<DefaultBuildHasher> {
+impl FzStringSet<Box<str>, DefaultBuildHasher> {
     /// Creates a new frozen set.
     #[doc = include_str!("../doc_snippets/duplicate_values.md")]
     #[must_use]
@@ -42,7 +43,7 @@ impl FzStringSet<DefaultBuildHasher> {
     }
 }
 
-impl<BH> FzStringSet<BH> {
+impl<BH> FzStringSet<Box<str>, BH> {
     /// Creates a new frozen set which uses the given hash builder to hash values.
     #[doc = include_str!("../doc_snippets/duplicate_values.md")]
     #[must_use]
@@ -54,13 +55,38 @@ impl<BH> FzStringSet<BH> {
             map: FzStringMap::with_hasher(entries.into_iter().map(|x| (x, ())).collect(), bh),
         }
     }
+}
 
+impl<'a> FzStringSet<&'a str, DefaultBuildHasher> {
+    /// Creates a new frozen set.
+    #[doc = include_str!("../doc_snippets/duplicate_values.md")]
+    #[must_use]
+    pub fn new_for_str(entries: Vec<&'a str>) -> Self {
+        Self::with_hasher_for_str(entries, DefaultBuildHasher::default())
+    }
+}
+
+impl<'a, BH> FzStringSet<&'a str, BH> {
+    /// Creates a new frozen set which uses the given hash builder to hash values.
+    #[doc = include_str!("../doc_snippets/duplicate_values.md")]
+    #[must_use]
+    pub fn with_hasher_for_str(entries: Vec<&'a str>, bh: BH) -> Self
+    where
+        BH: BuildHasher + Clone + Send + Sync + 'static,
+    {
+        Self {
+            map: FzStringMap::with_hasher_for_str(entries.into_iter().map(|x| (x, ())).collect(), bh),
+        }
+    }
+}
+
+impl<T, BH> FzStringSet<T, BH> {
     #[doc = include_str!("../doc_snippets/get_from_set.md")]
     #[inline]
-    #[expect(clippy::borrowed_box, reason = "Key type is Box<str>")]
-    pub fn get(&self, value: impl AsRef<str>) -> Option<&Box<str>>
+    pub fn get(&self, value: impl AsRef<str>) -> Option<&T>
     where
         BH: BuildHasher,
+        str: Equivalent<T>,
     {
         Some(self.map.get_key_value(value)?.0)
     }
@@ -71,6 +97,7 @@ impl<BH> FzStringSet<BH> {
     pub fn contains(&self, value: impl AsRef<str>) -> bool
     where
         BH: BuildHasher,
+        str: Equivalent<T>,
     {
         self.map.contains_key(value)
     }
@@ -91,16 +118,16 @@ impl<BH> FzStringSet<BH> {
 
     #[doc = include_str!("../doc_snippets/iter.md")]
     #[must_use]
-    pub fn iter(&self) -> Iter<'_, Box<str>> {
+    pub fn iter(&self) -> Iter<'_, T> {
         Iter::new(self.map.iter())
     }
 
-    fn into_iter(self) -> IntoIter<Box<str>> {
+    fn into_iter(self) -> IntoIter<T> {
         IntoIter::new(self.map.into_iter())
     }
 }
 
-impl<BH> Default for FzStringSet<BH>
+impl<BH> Default for FzStringSet<Box<str>, BH>
 where
     BH: Default,
 {
@@ -111,13 +138,24 @@ where
     }
 }
 
-impl<BH> From<FzStringMap<(), BH>> for FzStringSet<BH> {
-    fn from(map: FzStringMap<(), BH>) -> Self {
+impl<BH> Default for FzStringSet<&str, BH>
+where
+    BH: Default,
+{
+    fn default() -> Self {
+        Self {
+            map: FzStringMap::default(),
+        }
+    }
+}
+
+impl<T, BH> From<FzStringMap<T, (), BH>> for FzStringSet<T, BH> {
+    fn from(map: FzStringMap<T, (), BH>) -> Self {
         Self { map }
     }
 }
 
-impl<T, const N: usize, BH> From<[T; N]> for FzStringSet<BH>
+impl<T, const N: usize, BH> From<[T; N]> for FzStringSet<Box<str>, BH>
 where
     T: AsRef<str>,
     BH: BuildHasher + Default + Clone + Send + Sync + 'static,
@@ -127,7 +165,16 @@ where
     }
 }
 
-impl<T, BH> FromIterator<T> for FzStringSet<BH>
+impl<'a, const N: usize, BH> From<[&'a str; N]> for FzStringSet<&'a str, BH>
+where
+    BH: BuildHasher + Default + Clone + Send + Sync + 'static,
+{
+    fn from(entries: [&'a str; N]) -> Self {
+        Self::from(FzStringMap::from_iter(entries.into_iter().map(|x| (x, ()))))
+    }
+}
+
+impl<T, BH> FromIterator<T> for FzStringSet<Box<str>, BH>
 where
     T: AsRef<str>,
     BH: BuildHasher + Default + Clone + Send + Sync + 'static,
@@ -137,134 +184,167 @@ where
     }
 }
 
-impl<Q, BH> Set<Box<str>, Q> for FzStringSet<BH>
+impl<'a, BH> FromIterator<&'a str> for FzStringSet<&'a str, BH>
+where
+    BH: BuildHasher + Default + Clone + Send + Sync + 'static,
+{
+    fn from_iter<IT: IntoIterator<Item = &'a str>>(iter: IT) -> Self {
+        Self::from(FzStringMap::from_iter(iter.into_iter().map(|x| (x, ()))))
+    }
+}
+
+impl<T, Q, BH> Set<T, Q> for FzStringSet<T, BH>
 where
     Q: AsRef<str>,
     BH: BuildHasher,
+    str: Equivalent<T>,
 {
 }
 
-impl<Q, BH> SetExtras<Box<str>, Q> for FzStringSet<BH>
+impl<T, Q, BH> SetExtras<T, Q> for FzStringSet<T, BH>
 where
     Q: AsRef<str>,
     BH: BuildHasher,
+    str: Equivalent<T>,
 {
     #[inline]
-    fn get(&self, value: &Q) -> Option<&Box<str>> {
+    fn get(&self, value: &Q) -> Option<&T> {
         self.get(value)
     }
 }
 
-impl<Q, BH> SetQuery<Q> for FzStringSet<BH>
+impl<T, Q, BH> SetQuery<Q> for FzStringSet<T, BH>
 where
     Q: AsRef<str>,
     BH: BuildHasher,
+    str: Equivalent<T>,
 {
     set_query_trait_funcs!();
 }
 
-impl<BH> SetIteration<Box<str>> for FzStringSet<BH> {
+impl<T, BH> SetIteration<T> for FzStringSet<T, BH> {
     type Iterator<'a>
-        = Iter<'a, Box<str>>
+        = Iter<'a, T>
     where
+        T: 'a,
         BH: 'a;
 
-    fn iter(&self) -> Iter<'_, Box<str>> {
+    fn iter(&self) -> Iter<'_, T> {
         self.iter()
     }
 }
 
-impl<BH> Len for FzStringSet<BH> {
+impl<T, BH> Len for FzStringSet<T, BH> {
     len_trait_funcs!();
 }
 
-impl<ST, BH> BitOr<&ST> for &FzStringSet<BH>
+impl<T, ST, BH> BitOr<&ST> for &FzStringSet<T, BH>
 where
-    ST: Set<Box<str>>,
+    T: Hash + Eq + Clone + AsRef<str>,
+    ST: Set<T>,
     BH: BuildHasher + Default,
+    str: Equivalent<T>,
 {
-    type Output = hashbrown::HashSet<Box<str>>;
+    type Output = hashbrown::HashSet<T>;
 
     fn bitor(self, rhs: &ST) -> Self::Output {
         Self::Output::from_iter(self.union(rhs).cloned())
     }
 }
 
-impl<ST, BH> BitAnd<&ST> for &FzStringSet<BH>
+impl<T, ST, BH> BitAnd<&ST> for &FzStringSet<T, BH>
 where
-    ST: Set<Box<str>>,
+    T: Hash + Eq + Clone + AsRef<str>,
+    ST: Set<T>,
     BH: BuildHasher + Default,
+    str: Equivalent<T>,
 {
-    type Output = hashbrown::HashSet<Box<str>>;
+    type Output = hashbrown::HashSet<T>;
 
     fn bitand(self, rhs: &ST) -> Self::Output {
         Self::Output::from_iter(self.intersection(rhs).cloned())
     }
 }
 
-impl<ST, BH> BitXor<&ST> for &FzStringSet<BH>
+impl<T, ST, BH> BitXor<&ST> for &FzStringSet<T, BH>
 where
-    ST: Set<Box<str>>,
+    T: Hash + Eq + Clone + AsRef<str>,
+    ST: Set<T>,
     BH: BuildHasher + Default,
+    str: Equivalent<T>,
 {
-    type Output = hashbrown::HashSet<Box<str>>;
+    type Output = hashbrown::HashSet<T>;
 
     fn bitxor(self, rhs: &ST) -> Self::Output {
         self.symmetric_difference(rhs).cloned().collect()
     }
 }
 
-impl<ST, BH> Sub<&ST> for &FzStringSet<BH>
+impl<T, ST, BH> Sub<&ST> for &FzStringSet<T, BH>
 where
-    ST: Set<Box<str>>,
+    T: Hash + Eq + Clone + AsRef<str>,
+    ST: Set<T>,
     BH: BuildHasher + Default,
+    str: Equivalent<T>,
 {
-    type Output = hashbrown::HashSet<Box<str>>;
+    type Output = hashbrown::HashSet<T>;
 
     fn sub(self, rhs: &ST) -> Self::Output {
         self.difference(rhs).cloned().collect()
     }
 }
 
-impl<BH> IntoIterator for FzStringSet<BH> {
-    type Item = Box<str>;
-    type IntoIter = IntoIter<Box<str>>;
+impl<T, BH> IntoIterator for FzStringSet<T, BH> {
+    type Item = T;
+    type IntoIter = IntoIter<T>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.into_iter()
     }
 }
 
-impl<'a, BH> IntoIterator for &'a FzStringSet<BH> {
-    type Item = &'a Box<str>;
-    type IntoIter = Iter<'a, Box<str>>;
+impl<'a, T, BH> IntoIterator for &'a FzStringSet<T, BH> {
+    type Item = &'a T;
+    type IntoIter = Iter<'a, T>;
 
     fn into_iter(self) -> Self::IntoIter {
         self.iter()
     }
 }
 
-impl<ST, BH> PartialEq<ST> for FzStringSet<BH>
+impl<T, ST, BH> PartialEq<ST> for FzStringSet<T, BH>
 where
-    ST: SetQuery<Box<str>>,
+    ST: SetQuery<T>,
     BH: BuildHasher + Default,
 {
     partial_eq_trait_funcs!();
 }
 
-impl<BH> Eq for FzStringSet<BH> where BH: BuildHasher + Default {}
+impl<T, BH> Eq for FzStringSet<T, BH>
+where
+    T: AsRef<str>,
+    BH: BuildHasher + Default,
+    str: Equivalent<T>,
+{
+}
 
-impl<BH> Debug for FzStringSet<BH> {
+impl<T, BH> Debug for FzStringSet<T, BH>
+where
+    T: Debug,
+{
     debug_trait_funcs!();
 }
 
 #[cfg(feature = "serde")]
-impl Serialize for FzStringSet {
+impl<T, BH> Serialize for FzStringSet<T, BH>
+where
+    T: Serialize,
+{
     serialize_trait_funcs!();
 }
 
 #[cfg(feature = "serde")]
-impl<'de, BH> Deserialize<'de> for FzStringSet<BH>
+impl<'de, BH> Deserialize<'de> for FzStringSet<Box<str>, BH>
 where
     BH: BuildHasher + Default + Clone + Send + Sync + 'static,
 {
@@ -286,7 +366,7 @@ impl<'de, BH> Visitor<'de> for StrSetVisitor<BH>
 where
     BH: BuildHasher + Default + Clone + Send + Sync + 'static,
 {
-    type Value = FzStringSet<BH>;
+    type Value = FzStringSet<Box<str>, BH>;
 
     fn expecting(&self, formatter: &mut Formatter) -> core::fmt::Result {
         formatter.write_str("a set with string values")